requirements.txt
ipdb==0.13.13
Django==5.1.4
taskipy==1.13.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
dj-database-url==2.3.0
whitenoise[brotli]==6.8.2
gunicorn==23.0.0
uvicorn==0.32.0
supabase==2.15.2
